Be consistent about data types to fix few compiler warnings.
- Add cast to DWORD type when needed. Optionaly GrowableArray API could be
  changed to use size_t but this would require changes across the
  codebase to fix all usages.

+++ demuxer/Demuxers/LAVFDemuxer.cpp
@@ -2092,13 +2092,13 @@ STDMETHODIMP CLAVFDemuxer::CreateStreams()
       pg_sequences.sort();
       pg_sequences.unique();
 
-      int size = pg_sequences.size() * 4;
+      size_t size = pg_sequences.size() * 4;
       char *offsets = new char[size];
       offsets[0] = 0;
 
       // Append all offsets to the string
       for (auto it = pg_sequences.begin(); it != pg_sequences.end(); it++) {
-        int len = strlen(offsets);
+        size_t len = strlen(offsets);
         if (len > 0) {
           offsets[len] = ',';
           len++;
